% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianReasoning-package.R
\docType{package}
\name{BayesianReasoning-package}
\alias{BayesianReasoning}
\alias{BayesianReasoning-package}
\title{BayesianReasoning: Plot Positive and Negative Predictive Values for Medical Tests}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions to plot and help understand positive and negative predictive values (PPV and NPV), and their relationship with sensitivity, specificity, and prevalence. See Akobeng, A.K. (2007) \doi{10.1111/j.1651-2227.2006.00180.x} for a theoretical overview of the technical concepts and Navarrete et al. (2015) for a practical explanation about the importance of their understanding \doi{10.3389/fpsyg.2015.01327}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gorkang/BayesianReasoning}
  \item Report bugs at \url{https://github.com/gorkang/BayesianReasoning/issues}
}

}
\author{
\strong{Maintainer}: Gorka Navarrete \email{gorkang@gmail.com} (\href{https://orcid.org/0000-0001-7678-8656}{ORCID})

}
\keyword{internal}
