% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.createPPVmatrix}
\alias{.createPPVmatrix}
\title{.createPPVmatrix}
\usage{
.createPPVmatrix(
  Min_Prevalence = 1,
  Max_Prevalence,
  Sensitivity,
  Max_FP,
  PPV_NPV = "PPV"
)
}
\arguments{
\item{Max_Prevalence}{Max prevalence of disease}

\item{Sensitivity}{Sensitivity of test}

\item{Max_FP}{Maximum False Positives ratio}
}
\value{
A DF called PPV
}
\description{
Create a PPV matrix helper function
}
\examples{
.createPPVmatrix(1, 1000, 100, 2)
}
