% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.get_point_ppv_npv}
\alias{.get_point_ppv_npv}
\title{Get point ppv-npv}
\usage{
.get_point_ppv_npv(
  PPV_melted,
  PPV_NPV = "PPV",
  Sensitivity,
  overlay_prevalence_1,
  overlay_prevalence_2,
  overlay_labels,
  overlay_position_FP_FN,
  point_Prevalence
)
}
\arguments{
\item{overlay_position_FP_FN}{}
}
\value{

}
\description{
Get point ppv-npv
}
\examples{
.get_point_ppv_npv(PPV_melted, PPV_NPV = 'PPV', overlay_position_FP_FN = 1)
}
