% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.plot_creation}
\alias{.plot_creation}
\title{.plot_creation}
\usage{
.plot_creation(
  PPV_melted,
  min_Prevalence,
  max_Prevalence,
  PPV_NPV = "PPV",
  min_FP = 0,
  max_FP,
  max_FN,
  min_FN,
  one_out_of = TRUE,
  steps_matrix = 100,
  decimals_x,
  decimals_y,
  label_title = "",
  label_subtitle = "",
  translated_labels = translated_labels,
  DEBUG_MESSAGE = ""
)
}
\arguments{
\item{PPV_melted}{.}

\item{min_Prevalence}{.}

\item{PPV_NPV}{.}

\item{min_FP}{.}

\item{max_FP}{.}

\item{steps_matrix}{.}

\item{decimals_x}{.}

\item{decimals_y}{.}

\item{label_title}{.}

\item{label_subtitle}{.}

\item{translated_labels}{.}

\item{Sensitivity}{.}
}
\description{
Function to create the main heatmap plot
}
