% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.plot_overlay_area}
\alias{.plot_overlay_area}
\title{.plot_overlay_area}
\usage{
.plot_overlay_area(
  PPV_melted,
  uncertainty_prevalence = "low",
  min_Prevalence,
  max_Prevalence,
  Sensitivity,
  Specificity,
  min_FP = 0,
  max_FP,
  max_FN,
  min_FN,
  one_out_of = TRUE,
  overlay_labels = "",
  PPV_NPV = "PPV",
  Language = "en",
  overlay_prevalence_1,
  overlay_prevalence_2,
  overlay_position_FP,
  overlay_position_FN,
  decimals_x,
  decimals_y,
  label_title,
  label_subtitle,
  translated_labels = translated_labels
)
}
\arguments{
\item{PPV_melted}{.}

\item{uncertainty_prevalence}{.}

\item{min_Prevalence}{[x] out of y prevalence of disease}

\item{max_Prevalence}{x out of [y] prevalence of disease}

\item{Sensitivity}{.}

\item{min_FP}{.}

\item{max_FP}{.}

\item{overlay_labels}{.}

\item{PPV_NPV}{.}

\item{Language}{.}

\item{overlay_prevalence_1}{[x] out of y prevalence of disease for the overlay}

\item{overlay_prevalence_2}{x out of [y] prevalence of disease for the overlay}

\item{overlay_position_FP}{.}

\item{overlay_position_FN}{.}

\item{decimals_x}{.}

\item{decimals_y}{.}

\item{label_title}{.}

\item{label_subtitle}{.}

\item{translated_labels}{.}
}
\description{
Add area overlay to PPV_heatmap plot
}
