% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.translate_labels}
\alias{.translate_labels}
\title{.translate_labels}
\usage{
.translate_labels(Language, Sensitivity, Max_FP)
}
\arguments{
\item{Language}{}

\item{Sensitivity}{}

\item{Max_FP}{}
}
\value{

}
\description{
.translate_labels
}
\examples{
.translate_labels <- function(Language = "sp", Sensitivity = 1, Max_FP = 1)
}
