% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{process_variables}
\alias{process_variables}
\title{process_variables
Process main variables, checks for errors, creates ranges...}
\usage{
process_variables(
  Sensitivity,
  Specificity,
  width_Sensitivity = 10,
  width_Specificity = 10
)
}
\arguments{
\item{Sensitivity}{}

\item{Specificity}{}

\item{width_Sensitivity}{}

\item{width_Specificity}{}
}
\description{
process_variables
Process main variables, checks for errors, creates ranges...
}
