% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.createPPVmatrix}
\alias{.createPPVmatrix}
\title{.createPPVmatrix}
\usage{
.createPPVmatrix(
  Min_Prevalence = 1,
  Max_Prevalence = 1000,
  Sensitivity = 100,
  Min_FP = 0,
  Max_FP = 10,
  PPV_NPV = "PPV",
  steps_matrix = 100
)
}
\arguments{
\item{Min_Prevalence}{[x] out of y prevalence of disease}

\item{Max_Prevalence}{x out of [y] prevalence of disease}

\item{Sensitivity}{Sensitivity of test}

\item{Min_FP}{.}

\item{Max_FP}{Maximum False Positives ratio}

\item{PPV_NPV}{.}

\item{steps_matrix}{.}
}
\value{
A DF called PPV
}
\description{
Create a PPV matrix helper function
}
