% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.createPPVmatrix}
\alias{.createPPVmatrix}
\title{.createPPVmatrix}
\usage{
.createPPVmatrix(
  min_Prevalence = 1,
  max_Prevalence = 1000,
  Sensitivity = 100,
  Specificity = 99,
  one_out_of = TRUE,
  PPV_NPV = "PPV",
  min_FP = 0,
  max_FP = 10,
  max_FN = 0,
  min_FN = 10,
  steps_matrix = 100
)
}
\arguments{
\item{min_Prevalence}{[x] out of y prevalence of disease}

\item{max_Prevalence}{x out of [y] prevalence of disease}

\item{Sensitivity}{Sensitivity of test}

\item{Specificity}{Specificity of test}

\item{one_out_of}{.}

\item{PPV_NPV}{Should show PPV or NPV ["PPV", "NPV"]}

\item{min_FP}{Minimum False Positive ratio}

\item{max_FP}{Maximum False Positive ratio}

\item{max_FN}{.}

\item{min_FN}{.}

\item{steps_matrix}{How big the matrix should be (probably better to leave as it is: 100)}
}
\value{
A DF called PPV_melted
}
\description{
Create a PPV matrix helper function
}
