% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.get_point_ppv_npv}
\alias{.get_point_ppv_npv}
\title{.get_point_ppv_npv}
\usage{
.get_point_ppv_npv(
  PPV_melted,
  PPV_NPV = "PPV",
  Language,
  Sensitivity,
  Specificity,
  overlay_prevalence_1,
  overlay_prevalence_2,
  overlay_labels,
  overlay_extra_info = FALSE,
  overlay_position_FP,
  overlay_position_FN,
  point_Prevalence,
  translated_labels,
  decimals_x,
  decimals_y
)
}
\arguments{
\item{PPV_melted}{DF out of .createPPVmatrix()}

\item{PPV_NPV}{Should calculate PPV or NPV?}

\item{Language}{["en" / "es"]}

\item{Sensitivity}{Sensitivity of the test}

\item{Specificity}{Specificity of the test}

\item{overlay_prevalence_1}{[x] out of y prevalence of disease}

\item{overlay_prevalence_2}{x out of [y] prevalence of disease}

\item{overlay_labels}{vector with labels for each overlay point}

\item{overlay_extra_info}{show extra info in overlay? [TRUE/FALSE]}

\item{overlay_position_FP}{.}

\item{overlay_position_FN}{.}

\item{point_Prevalence}{.}

\item{translated_labels}{.}

\item{decimals_x}{.}

\item{decimals_y}{.}
}
\description{
Get PPV or NPV for the overlay
}
