% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.get_point_ppv_npv}
\alias{.get_point_ppv_npv}
\title{.get_point_ppv_npv}
\usage{
.get_point_ppv_npv(
  PPV_melted,
  PPV_NPV = "PPV",
  Sensitivity,
  overlay_prevalence_1,
  overlay_prevalence_2,
  overlay_labels,
  overlay_position_FP,
  overlay_position_FN,
  point_Prevalence,
  prevalence_label,
  x_axis_label,
  y_axis_label,
  decimals_x,
  decimals_y
)
}
\arguments{
\item{PPV_melted}{.}

\item{PPV_NPV}{.}

\item{Sensitivity}{.}

\item{overlay_prevalence_1}{.}

\item{overlay_prevalence_2}{.}

\item{overlay_labels}{.}

\item{overlay_position_FP}{.}

\item{overlay_position_FN}{.}

\item{point_Prevalence}{.}

\item{prevalence_label}{.}

\item{x_axis_label}{.}

\item{y_axis_label}{.}

\item{decimals_x}{.}

\item{decimals_y}{.}
}
\description{
Description
}
