% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.number_decimals_plot_axis}
\alias{.number_decimals_plot_axis}
\title{.number_decimals_plot_axis}
\usage{
.number_decimals_plot_axis(
  PPV_NPV = "PPV",
  min_FP = 0,
  max_FP,
  min_FN,
  max_FN,
  min_Prevalence,
  max_Prevalence
)
}
\arguments{
\item{PPV_NPV}{.}

\item{min_FP}{.}

\item{max_FP}{.}

\item{min_FN}{.}

\item{max_FN}{.}

\item{min_Prevalence}{[x] out of y prevalence of disease}

\item{max_Prevalence}{x out of [y] prevalence of disease}
}
\description{
The number of decimal places in the x and y axis label depends on how wide the range is.
}
