% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.number_decimals_plot_axis}
\alias{.number_decimals_plot_axis}
\title{.number_decimals_plot_axis}
\usage{
.number_decimals_plot_axis(
  PPV_NPV = "PPV",
  Min_FP = 0,
  Max_FP,
  Min_FN,
  Max_FN,
  Min_Prevalence,
  Max_Prevalence
)
}
\arguments{
\item{PPV_NPV}{.}

\item{Min_FP}{.}

\item{Max_FP}{.}

\item{Min_FN}{.}

\item{Max_FN}{.}

\item{Min_Prevalence}{.}

\item{Max_Prevalence}{.}
}
\description{
.number_decimals_plot_axis
}
