% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.plot_creation}
\alias{.plot_creation}
\title{.plot_creation}
\usage{
.plot_creation(
  PPV_melted,
  Sensitivity,
  PPV_NPV = "PPV",
  Min_Prevalence = 1,
  Min_FP = 0,
  Max_FP,
  label_title = "",
  label_subtitle = "",
  label_caption = "",
  prevalence_label,
  steps_matrix = 100,
  decimals_x,
  decimals_y,
  legend_label = "VPP (\%)\\n"
)
}
\arguments{
\item{PPV_melted}{}

\item{label_title}{}

\item{label_subtitle}{}
}
\value{

}
\description{
.plot_creation
}
