% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.plot_creation}
\alias{.plot_creation}
\title{.plot_creation}
\usage{
.plot_creation(
  PPV_melted,
  min_Prevalence,
  max_Prevalence,
  Sensitivity,
  PPV_NPV = "PPV",
  min_FP = 0,
  max_FP,
  Max_FN,
  Min_FN,
  one_out_of = TRUE,
  steps_matrix = 100,
  decimals_x,
  decimals_y,
  label_title = "",
  label_subtitle = "",
  label_caption = "",
  prevalence_label = "",
  legend_label = "",
  x_axis_label,
  x_axis_label_extra,
  y_axis_label,
  DEBUG_MESSAGE = ""
)
}
\arguments{
\item{PPV_melted}{.}

\item{min_Prevalence}{.}

\item{Sensitivity}{.}

\item{PPV_NPV}{.}

\item{min_FP}{.}

\item{max_FP}{.}

\item{steps_matrix}{.}

\item{decimals_x}{.}

\item{decimals_y}{.}

\item{label_title}{.}

\item{label_subtitle}{.}

\item{label_caption}{.}

\item{prevalence_label}{.}

\item{legend_label}{.}

\item{x_axis_label}{.}

\item{x_axis_label_extra}{.}

\item{y_axis_label}{.}
}
\description{
Function to create the main heatmap plot
}
