% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.plot_creation}
\alias{.plot_creation}
\title{.plot_creation}
\usage{
.plot_creation(
  PPV_melted,
  Max_FP,
  Step_size_FP,
  label_title = "",
  label_subtitle = "",
  label_caption = "",
  decimals_x,
  decimals_y,
  prevalence_label
)
}
\arguments{
\item{PPV_melted}{}

\item{label_title}{}

\item{label_subtitle}{}
}
\value{

}
\description{
.plot_creation
}
