% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.plot_creation}
\alias{.plot_creation}
\title{.plot_creation}
\usage{
.plot_creation(
  PPV_melted,
  Min_Prevalence,
  Sensitivity,
  PPV_NPV = "PPV",
  Min_FP = 0,
  Max_FP,
  steps_matrix = 100,
  decimals_x,
  decimals_y,
  label_title = "",
  label_subtitle = "",
  label_caption = "",
  prevalence_label = "",
  legend_label = "",
  x_axis_label,
  y_axis_label
)
}
\arguments{
\item{PPV_melted}{.}

\item{Min_Prevalence}{.}

\item{Sensitivity}{.}

\item{PPV_NPV}{.}

\item{Min_FP}{.}

\item{Max_FP}{.}

\item{steps_matrix}{.}

\item{decimals_x}{.}

\item{decimals_y}{.}

\item{label_title}{.}

\item{label_subtitle}{.}

\item{label_caption}{.}

\item{prevalence_label}{.}

\item{legend_label}{.}

\item{x_axis_label}{.}

\item{y_axis_label}{.}
}
\description{
Function to create the main heatmap plot
}
