% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.plot_overlay_area}
\alias{.plot_overlay_area}
\title{Plot overlay area}
\usage{
.plot_overlay_area(
  PPV_melted,
  uncertainty_prevalence = "low",
  Min_Prevalence,
  Max_Prevalence,
  Sensitivity,
  Max_FP,
  overlay_labels = "",
  PPV_NPV = "PPV",
  overlay_prevalence_1,
  overlay_prevalence_2,
  overlay_position_FP_FN,
  decimals_x,
  decimals_y,
  prevalence_label
)
}
\arguments{
\item{PPV_melted}{}

\item{uncertainty_prevalence}{}

\item{overlay_labels}{}

\item{PPV_NPV}{}

\item{overlay_position_FP_FN}{}
}
\value{

}
\description{
Plot overlay area
}
