% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.plot_overlay_area}
\alias{.plot_overlay_area}
\title{.plot_overlay_area}
\usage{
.plot_overlay_area(
  PPV_NPV = "PPV",
  one_out_of = TRUE,
  min_Prevalence,
  max_Prevalence,
  min_FP = 0,
  max_FP,
  max_FN,
  min_FN,
  PPV_melted,
  steps_matrix = 100,
  decimals_x,
  decimals_y,
  label_title = "",
  label_subtitle = "",
  translated_labels = translated_labels,
  Language = "en",
  Sensitivity,
  Specificity,
  uncertainty_prevalence = "low",
  overlay_prevalence_1,
  overlay_prevalence_2,
  overlay_position_FP,
  overlay_position_FN,
  overlay_labels = "",
  overlay_extra_info = FALSE,
  ...
)
}
\arguments{
\item{PPV_NPV}{.}

\item{one_out_of}{.}

\item{min_Prevalence}{[x] out of y prevalence of disease}

\item{max_Prevalence}{x out of [y] prevalence of disease}

\item{min_FP}{.}

\item{max_FP}{.}

\item{max_FN}{.}

\item{min_FN}{.}

\item{PPV_melted}{.}

\item{steps_matrix}{.}

\item{decimals_x}{.}

\item{decimals_y}{.}

\item{label_title}{.}

\item{label_subtitle}{.}

\item{translated_labels}{.}

\item{Language}{.}

\item{Sensitivity}{.}

\item{Specificity}{.}

\item{uncertainty_prevalence}{.}

\item{overlay_prevalence_1}{[x] out of y prevalence of disease for the overlay}

\item{overlay_prevalence_2}{x out of [y] prevalence of disease for the overlay}

\item{overlay_position_FP}{.}

\item{overlay_position_FN}{.}

\item{overlay_labels}{vector with labels for each overlay point}

\item{overlay_extra_info}{show extra info in overlay? [TRUE/FALSE]}

\item{...}{.}
}
\description{
Add area overlay to PPV_heatmap plot
}
