% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.plot_overlay_area}
\alias{.plot_overlay_area}
\title{.plot_overlay_area}
\usage{
.plot_overlay_area(
  PPV_melted,
  uncertainty_prevalence = "low",
  Min_Prevalence,
  Max_Prevalence,
  Sensitivity,
  Min_FP = 0,
  Max_FP,
  overlay_labels = "",
  PPV_NPV = "PPV",
  overlay_prevalence_1,
  overlay_prevalence_2,
  overlay_position_FP,
  overlay_position_FN,
  decimals_x,
  decimals_y,
  prevalence_label,
  legend_label,
  label_title,
  label_subtitle,
  x_axis_label,
  y_axis_label
)
}
\arguments{
\item{PPV_melted}{.}

\item{uncertainty_prevalence}{.}

\item{Min_Prevalence}{.}

\item{Max_Prevalence}{.}

\item{Sensitivity}{.}

\item{Min_FP}{.}

\item{Max_FP}{.}

\item{overlay_labels}{.}

\item{PPV_NPV}{.}

\item{overlay_prevalence_1}{.}

\item{overlay_prevalence_2}{.}

\item{overlay_position_FP}{.}

\item{overlay_position_FN}{.}

\item{decimals_x}{.}

\item{decimals_y}{.}

\item{prevalence_label}{.}

\item{legend_label}{.}

\item{label_title}{.}

\item{label_subtitle}{.}

\item{x_axis_label}{.}

\item{y_axis_label}{.}
}
\description{
Description
}
