% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.plot_overlay_line}
\alias{.plot_overlay_line}
\title{.plot_overlay_line}
\usage{
.plot_overlay_line(
  PPV_melted,
  uncertainty_prevalence = "low",
  PPV_NPV = "PPV",
  min_Prevalence,
  max_Prevalence,
  min_FP = 0,
  max_FP,
  max_FN,
  min_FN,
  one_out_of = TRUE,
  overlay_prevalence_1,
  overlay_prevalence_2,
  overlay_position_FP,
  overlay_position_FN,
  decimals_x,
  decimals_y,
  overlay_labels,
  label_title,
  label_subtitle,
  translated_labels = translated_labels
)
}
\arguments{
\item{PPV_melted}{DF}

\item{uncertainty_prevalence}{How big the uncertainty area should be: ["low" or "high"]}

\item{PPV_NPV}{.}

\item{min_Prevalence}{[x] out of y prevalence of disease}

\item{max_Prevalence}{x out of [y] prevalence of disease}

\item{min_FP}{}

\item{max_FP}{.}

\item{max_FN}{}

\item{min_FN}{}

\item{one_out_of}{}

\item{overlay_prevalence_1}{vector with [x] out of y prevalence of disease}

\item{overlay_prevalence_2}{vector with x out of [y] prevalence of disease}

\item{overlay_position_FP}{.}

\item{overlay_position_FN}{.}

\item{decimals_x}{.}

\item{decimals_y}{.}

\item{overlay_labels}{vector with labels for each overlay point}

\item{label_title}{.}

\item{label_subtitle}{.}

\item{translated_labels}{.}

\item{Sensitivity}{.}

\item{Specificity}{}
}
\description{
Add line overlay to PPV_heatmap plot
}
