% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.translate_labels}
\alias{.translate_labels}
\title{.translate_labels}
\usage{
.translate_labels(Language, Sensitivity, Max_FP, PPV_NPV = "PPV")
}
\arguments{
\item{Language}{.}

\item{Sensitivity}{.}

\item{Max_FP}{.}

\item{PPV_NPV}{.}
}
\value{
A list with labels
}
\description{
.translate_labels
}
