% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_possible_prevalence.R
\name{min_possible_prevalence}
\alias{min_possible_prevalence}
\title{Show minimum possible prevalence}
\usage{
min_possible_prevalence(Sensitivity, FP_test, min_PPV_desired)
}
\arguments{
\item{Sensitivity}{Sensitivity of the test.}

\item{FP_test}{False positive rate (1-Specificity) [0-100].}

\item{min_PPV_desired}{Which PPV is what you consider the minimum to trust a positive result in the test.}
}
\value{
A description showing the minimum necessary prevalence.
}
\description{
Given a FP and a desired PPV, what is the Minimum Prevalence of a Condition
}
\examples{

# Example 1
min_possible_prevalence(Sensitivity = 99.9, FP_test = 0.2, min_PPV_desired = 0.9)
> To get a PPV of 0.9 with a test with 99.9 \% Sensitivity and 0.2 \% False Positive Rate, you need a prevalence of at least 1 out of 56

# Example 2
min_possible_prevalence(100, 0.1, 0.98)
> To get a PPV of 0.98 with a test with 100 \% Sensitivity and 0.1 \% False Positive Rate, you need a prevalence of at least 1 out of 21
}
