% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPV_heatmap.R
\name{PPV_heatmap}
\alias{PPV_heatmap}
\title{Plot PPV heatmaps}
\usage{
PPV_heatmap(Max_Prevalence, Sensitivity, Max_FP, overlay = FALSE,
  overlay_labels, overlay_position_FP, overlay_position_Prevalence,
  label_title = "", label_subtitle = "", Language = "en",
  save_plot = TRUE)
}
\arguments{
\item{Max_Prevalence}{Maximum prevalence to show in plot (y-axis): 1-Inf}

\item{Sensitivity}{Sensitivity of the test: 0-100}

\item{Max_FP}{Maximum False Positives ratio to show in plot (x-axis): 1-100}

\item{overlay}{Show overlay: TRUE / FALSE}

\item{overlay_labels}{Lables for each point in the overlay. For example: c("80", "70", "60", "50", "40", "30", "20  y.o.")}

\item{overlay_position_FP}{FP value (position in the x-axis) for each point in the overlay. For example: c(7, 8, 9, 12, 14, 14)}

\item{overlay_position_Prevalence}{Prevalence value (position in the y-axis) for each point in the overlay. For example: c(26, 29, 44, 69, 227, 1667)}

\item{label_title}{Title for the plot}

\item{label_subtitle}{Subtitle for the plot}

\item{Language}{Language for the plot labels: "sp" / "en"}

\item{save_plot}{Should save the plot as .png or just show it?}
}
\value{
A .png plot in the /output folder, or shows a plot
}
\description{
Plot PPV heatmaps Plot heatmaps with PPV values for a given specificity and a range of Prevalences and FP
}
\examples{

# Simple plot
PPV_heatmap(Max_Prevalence = 500, Sensitivity = 90, Max_FP = 15, Language = "en", overlay = FALSE)

# Show a plot with an overlay
PPV_heatmap(Max_Prevalence = 1800, Sensitivity = 90, Max_FP = 15, 
               label_subtitle = "PPV of Mammogram for Breast Cancer by Age",
               save_plot = TRUE, Language = "en", 
                overlay = TRUE, 
                overlay_labels = c("80", "70", "60", "50", "40", "30", "20  y.o."),
                overlay_position_FP = c(7, 8, 9, 12, 14, 14),
                overlay_position_Prevalence = c(26, 29, 44, 69, 227, 1667))
}
