% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{process_variables}
\alias{process_variables}
\title{process_variables
Process main variables, checks for errors, creates ranges...}
\usage{
process_variables(
  Sensitivity,
  Specificity,
  limits_Sensitivity = NULL,
  limits_Specificity = NULL,
  PPV_NPV = "PPV"
)
}
\arguments{
\item{Sensitivity}{.}

\item{Specificity}{.}

\item{limits_Sensitivity}{.}

\item{limits_Specificity}{.}

\item{PPV_NPV}{.}
}
\description{
process_variables
Process main variables, checks for errors, creates ranges...
}
